+++ client/src/components/card/Card.jsx
@@ -8,7 +8,7 @@ export default function Card({name, img, id, price, type, color, handleAddCart})
 
   return (
     <div className={styles.card} >
-        <Link to={`/detail/${id}`}><p>{name}</p></Link>
+        <Link to={`/detail/${id}`}><h5>{name}</h5></Link>
         
         <div className={styles.imageContainer}>
           <img src={img} alt="imagen" />
