+++ client/src/pages/about/About.jsx
@@ -1,4 +1,5 @@
 import React from 'react'
+import Footer from '../../components/footer/Footer.jsx'
 import NavBar from '../../components/navBar/NavBar.jsx'
 
 export default function About() {
@@ -6,6 +7,7 @@ export default function About() {
     <div>
       <NavBar />
       <h1>About</h1>
+      <Footer/>
     </div>
   )
 }
