+++ client/src/pages/userAccount/UserAccount.module.css
@@ -1,0 +1,3 @@
+.UserAccComp {
+    min-height: 80vh;
+}
